requirements.txt
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
ARCtrl==2.0.0a2
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
EditorConfig==0.12.4
et-xmlfile==1.1.0
executing==2.0.1
Faker==24.7.1
fast-query-parsers==1.0.3
fsspreadsheet==6.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
icecream==2.1.3
idna==3.6
Jinja2==3.1.3
jsbeautifier==1.15.1
litestar==2.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgspec==0.18.6
multidict==6.0.5
openpyxl==3.1.2
polars==0.20.19
polyfactory==2.15.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.28.1
rich==13.7.1
rich-click==1.7.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
